feat: HQM-56 (#1199)
* add: bnjs types

* add: migrate from ethers to bnjs

* fix: migrate from formatters to dedeicated class

* add: types dependencies

* fix: code review

* add: balance type for staking sum

* fix: balance add method

* fix: active staking available balance

* update: Balance type and class

* add: CURRENCY_NAME

* add: toBalanceString method

+++ src/screens/staking-info.tsx
@@ -146,7 +146,7 @@ export const StakingInfoScreen = () => {
 
   const onDelegate = useCallback(async () => {
     const available = visible.filter(
-      v => app.getBalance(v.address) >= MIN_AMOUNT,
+      v => app.getBalance(v.address).toFloat() >= MIN_AMOUNT,
     );
 
     if (!available?.length) {
